backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
pydantic
faiss-cpu
pandas
python-dotenv
googletrans==4.0.0rc1
